% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurrentPath.R
\name{getCurrentPath}
\alias{getCurrentPath}
\title{Get the Current Script Path}
\usage{
getCurrentPath()
}
\value{
The normalized directory path of the script.
}
\description{
A function to determine the path of the currently running script.
}
